# [Proposal] Logs Bridge API - Logger.WithAttributes instead of Record.AddAttributes