--- a/log/benchmark/spy_test.go
+++ b/log/benchmark/spy_test.go
@@ -6,6 +6,7 @@ package benchmark
 import (
 	"context"
 
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/log/embedded"
 )
@@ -13,8 +14,14 @@ import (
 type spyLogger struct {
 	embedded.Logger
 	Record log.Record
+	Attrs  []attribute.KeyValue
 }
 
 func (l *spyLogger) Emit(_ context.Context, r log.Record) {
 	l.Record = r
+}
+
+func (l *spyLogger) WithAttributes(attrs ...attribute.KeyValue) log.Logger {
+	l.Attrs = attrs
+	return l
 }