--- a/log/go.mod
+++ b/log/go.mod
@@ -4,15 +4,8 @@ go 1.20
 
 require go.opentelemetry.io/otel v1.21.0
 
-require (
-	github.com/go-logr/logr v1.3.0 // indirect
-	github.com/go-logr/stdr v1.2.2 // indirect
-	go.opentelemetry.io/otel/metric v1.21.0 // indirect
-	go.opentelemetry.io/otel/trace v1.21.0 // indirect
-)
+replace go.opentelemetry.io/otel => ../
 
 replace go.opentelemetry.io/otel/trace => ../trace
 
 replace go.opentelemetry.io/otel/metric => ../metric
-
-replace go.opentelemetry.io/otel => ../