--- a/log/record.go
+++ b/log/record.go
@@ -10,9 +10,6 @@ package log // import "go.opentelemetry.io/otel/log"
 import (
 	"errors"
 	"time"
-
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
 )
 
 var errUnsafeAddAttrs = errors.New("unsafely called AddAttrs on copy of Record made without using Record.Clone")
@@ -20,31 +17,12 @@ var errUnsafeAddAttrs = errors.New("unsafely called AddAttrs on copy of Record m
 // Record TODO: comment.
 // TODO: Add unit tests.
 type Record struct {
-	timestamp         time.Time
-	observedTimestamp time.Time
-	severity          Severity
-	severityText      string
-	body              string
-
-	// The fields below are for optimizing the implementation of
-	// Attributes and AddAttributes.
-
-	// Allocation optimization: an inline array sized to hold
-	// the majority of log calls (based on examination of open-source
-	// code). It holds the start of the list of attributes.
-	front [attributesInlineCount]attribute.KeyValue
-
-	// The number of attributes in front.
-	nFront int
-
-	// The list of attributes except for those in front.
-	// Invariants:
-	//   - len(back) > 0 if nFront == len(front)
-	//   - Unused array elements are zero. Used to detect mistakes.
-	back []attribute.KeyValue
+	Timestamp         time.Time
+	ObservedTimestamp time.Time
+	Severity          Severity
+	SeverityText      string
+	Body              string
 }
-
-const attributesInlineCount = 5
 
 // Severity TODO: comment.
 type Severity int
@@ -77,146 +55,3 @@ const (
 	SeverityFatal3
 	SeverityFatal4
 )
-
-// Timestamp TODO: comment.
-func (r Record) Timestamp() time.Time {
-	return r.timestamp
-}
-
-// SetTimestamp TODO: comment.
-func (r *Record) SetTimestamp(t time.Time) {
-	r.timestamp = t
-}
-
-// ObservedTimestamp TODO: comment.
-func (r Record) ObservedTimestamp() time.Time {
-	return r.observedTimestamp
-}
-
-// SetObservedTimestamp TODO: comment.
-func (r *Record) SetObservedTimestamp(t time.Time) {
-	r.observedTimestamp = t
-}
-
-// Severity TODO: comment.
-func (r Record) Severity() Severity {
-	return r.severity
-}
-
-// SetSeverity TODO: comment.
-func (r *Record) SetSeverity(s Severity) {
-	r.severity = s
-}
-
-// SeverityText TODO: comment.
-func (r Record) SeverityText() string {
-	return r.severityText
-}
-
-// SetSeverityText TODO: comment.
-func (r *Record) SetSeverityText(s string) {
-	r.severityText = s
-}
-
-// Body TODO: comment.
-func (r Record) Body() string {
-	return r.body
-}
-
-// SetBody TODO: comment.
-func (r *Record) SetBody(s string) {
-	r.body = s
-}
-
-// WalkAttributes calls f on each [attribute.KeyValue] in the [Record].
-// Iteration stops if f returns false.
-func (r Record) WalkAttributes(f func(attribute.KeyValue) bool) {
-	for i := 0; i < r.nFront; i++ {
-		if !f(r.front[i]) {
-			return
-		}
-	}
-	for _, a := range r.back {
-		if !f(a) {
-			return
-		}
-	}
-}
-
-// AddAttributes appends the given [attribute.KeyValue] to the [Record]'s list of [attribute.KeyValue].
-// It omits invalid attributes.
-func (r *Record) AddAttributes(attrs ...attribute.KeyValue) {
-	var i int
-	for i = 0; i < len(attrs) && r.nFront < len(r.front); i++ {
-		a := attrs[i]
-		if !a.Valid() {
-			continue
-		}
-		r.front[r.nFront] = a
-		r.nFront++
-	}
-	// Check if a copy was modified by slicing past the end
-	// and seeing if the Attr there is non-zero.
-	if cap(r.back) > len(r.back) {
-		end := r.back[:len(r.back)+1][len(r.back)]
-		if end.Valid() {
-			// Don't panic; copy and muddle through.
-			r.back = sliceClip(r.back)
-			otel.Handle(errUnsafeAddAttrs)
-		}
-	}
-	ne := countInvalidAttrs(attrs[i:])
-	r.back = sliceGrow(r.back, len(attrs[i:])-ne)
-	for _, a := range attrs[i:] {
-		if a.Valid() {
-			r.back = append(r.back, a)
-		}
-	}
-}
-
-// Clone returns a copy of the record with no shared state.
-// The original record and the clone can both be modified
-// without interfering with each other.
-func (r Record) Clone() Record {
-	r.back = sliceClip(r.back) // prevent append from mutating shared array
-	return r
-}
-
-// AttributesLen returns the number of attributes in the Record.
-func (r Record) AttributesLen() int {
-	return r.nFront + len(r.back)
-}
-
-// countInvalidAttrs returns the number of invalid attributes.
-func countInvalidAttrs(as []attribute.KeyValue) int {
-	n := 0
-	for _, a := range as {
-		if !a.Valid() {
-			n++
-		}
-	}
-	return n
-}
-
-// sliceGrow increases the slice's capacity, if necessary, to guarantee space for
-// another n elements. After Grow(n), at least n elements can be appended
-// to the slice without another allocation. If n is negative or too large to
-// allocate the memory, Grow panics.
-//
-// This is a copy from https://pkg.go.dev/slices as it is not available in Go 1.20.
-func sliceGrow[S ~[]E, E any](s S, n int) S {
-	if n < 0 {
-		panic("cannot be negative")
-	}
-	if n -= cap(s) - len(s); n > 0 {
-		s = append(s[:cap(s)], make([]E, n)...)[:len(s)]
-	}
-	return s
-}
-
-// sliceClip removes unused capacity from the slice, returning s[:len(s):len(s)].
-//
-// This is a copy from https://pkg.go.dev/slices as it is not available in Go 1.20.
-func sliceClip[S ~[]E, E any](s S) S {
-	return s[:len(s):len(s)]
-}