Towards addressing https://github.com/open-telemetry/opentelemetry-go/pull/4725#discussion_r1419531687

Current results of `benchstat` compared to `logs-design` branch:

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/log/benchmark
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                        │   old.txt    │                new.txt                │
                        │    sec/op    │    sec/op      vs base                │
Emit/noop/no_attrs-16     24.43n ±  3%    10.47n ±  5%  -57.13% (p=0.000 n=10)
Emit/noop/3_attrs-16      73.69n ±  8%   132.60n ±  1%  +79.93% (p=0.000 n=10)
Emit/noop/5_attrs-16      106.9n ±  3%    178.2n ±  3%  +66.70% (p=0.000 n=10)
Emit/noop/10_attrs-16     378.8n ±  2%    299.4n ±  2%  -20.97% (p=0.000 n=10)
Emit/noop/40_attrs-16     1.324µ ±  2%    1.030µ ±  2%  -22.21% (p=0.000 n=10)
Emit/writer/no_attrs-16   204.6n ±  3%    142.2n ±  3%  -30.46% (p=0.000 n=10)
Emit/writer/3_attrs-16    616.8n ±  2%    748.9n ±  3%  +21.42% (p=0.000 n=10)
Emit/writer/5_attrs-16    755.4n ±  3%    890.6n ±  2%  +17.90% (p=0.000 n=10)
Emit/writer/10_attrs-16   1.576µ ±  5%    1.614µ ±  3%        ~ (p=0.075 n=10)
Emit/writer/40_attrs-16   5.513µ ± 24%    5.616µ ±  2%        ~ (p=0.353 n=10)
Slog/no_attrs-16          463.4n ±  3%    442.1n ±  5%   -4.61% (p=0.009 n=10)
Slog/3_attrs-16           706.8n ± 11%    978.5n ±  2%  +38.44% (p=0.000 n=10)
Slog/5_attrs-16           850.9n ± 13%   1255.5n ±  5%  +47.55% (p=0.000 n=10)
Slog/10_attrs-16          1.811µ ±  1%    2.110µ ±  2%  +16.54% (p=0.000 n=10)
Slog/40_attrs-16          6.267µ ±  3%    6.966µ ± 27%  +11.16% (p=0.000 n=10)
Logr/no_attrs-16          34.81n ±  2%    25.11n ±  8%  -27.88% (p=0.000 n=10)
Logr/3_attrs-16           262.7n ± 18%    425.9n ±  4%  +62.15% (p=0.000 n=10)
Logr/5_attrs-16           362.3n ±  4%    686.1n ±  3%  +89.35% (p=0.000 n=10)
Logr/10_attrs-16          1.115µ ±  9%    1.352µ ±  1%  +21.27% (p=0.000 n=10)
Logr/40_attrs-16          5.428µ ±  3%    5.352µ ±  6%        ~ (p=0.754 n=10)
geomean                   543.5n          591.8n         +8.88%

                        │    old.txt     │                 new.txt                  │
                        │      B/op      │     B/op      vs base                    │
Emit/noop/no_attrs-16       0.000 ± 0%       0.000 ± 0%          ~ (p=1.000 n=10) ¹
Emit/noop/3_attrs-16          0.0 ± 0%       208.0 ± 0%          ? (p=0.000 n=10)
Emit/noop/5_attrs-16          0.0 ± 0%       336.0 ± 0%          ? (p=0.000 n=10)
Emit/noop/10_attrs-16       320.0 ± 0%       656.0 ± 0%   +105.00% (p=0.000 n=10)
Emit/noop/40_attrs-16     2.250Ki ± 0%     2.641Ki ± 0%    +17.36% (p=0.000 n=10)
Emit/writer/no_attrs-16     16.00 ± 0%       16.00 ± 0%          ~ (p=1.000 n=10) ¹
Emit/writer/3_attrs-16      48.00 ± 0%      624.00 ± 0%  +1200.00% (p=0.000 n=10)
Emit/writer/5_attrs-16      48.00 ± 0%      752.00 ± 0%  +1466.67% (p=0.000 n=10)
Emit/writer/10_attrs-16     408.0 ± 0%      1432.0 ± 0%   +250.98% (p=0.000 n=10)
Emit/writer/40_attrs-16   2.570Ki ± 0%     5.570Ki ± 0%   +116.72% (p=0.000 n=10)
Slog/no_attrs-16            0.000 ± 0%       0.000 ± 0%          ~ (p=1.000 n=10) ¹
Slog/3_attrs-16               0.0 ± 0%       240.0 ± 0%          ? (p=0.000 n=10)
Slog/5_attrs-16               0.0 ± 0%       400.0 ± 0%          ? (p=0.000 n=10)
Slog/10_attrs-16           1168.0 ± 0%      1008.0 ± 0%    -13.70% (p=0.000 n=10)
Slog/40_attrs-16          9.312Ki ± 0%     4.500Ki ± 0%    -51.68% (p=0.000 n=10)
Logr/no_attrs-16            0.000 ± 0%       0.000 ± 0%          ~ (p=1.000 n=10) ¹
Logr/3_attrs-16             128.0 ± 0%       368.0 ± 0%   +187.50% (p=0.000 n=10)
Logr/5_attrs-16             208.0 ± 0%       608.0 ± 0%   +192.31% (p=0.000 n=10)
Logr/10_attrs-16          1.344Ki ± 0%     1.188Ki ± 0%    -11.63% (p=0.000 n=10)
Logr/40_attrs-16          9.562Ki ± 0%     4.750Ki ± 0%    -50.33% (p=0.000 n=10)
geomean                                ²                 ?                        ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                        │   old.txt    │                new.txt                 │
                        │  allocs/op   │  allocs/op   vs base                   │
Emit/noop/no_attrs-16     0.000 ± 0%      0.000 ± 0%         ~ (p=1.000 n=10) ¹
Emit/noop/3_attrs-16      0.000 ± 0%      2.000 ± 0%         ? (p=0.000 n=10)
Emit/noop/5_attrs-16      0.000 ± 0%      2.000 ± 0%         ? (p=0.000 n=10)
Emit/noop/10_attrs-16     1.000 ± 0%      2.000 ± 0%  +100.00% (p=0.000 n=10)
Emit/noop/40_attrs-16     1.000 ± 0%      2.000 ± 0%  +100.00% (p=0.000 n=10)
Emit/writer/no_attrs-16   1.000 ± 0%      1.000 ± 0%         ~ (p=1.000 n=10) ¹
Emit/writer/3_attrs-16    4.000 ± 0%      6.000 ± 0%   +50.00% (p=0.000 n=10)
Emit/writer/5_attrs-16    4.000 ± 0%      6.000 ± 0%   +50.00% (p=0.000 n=10)
Emit/writer/10_attrs-16   8.000 ± 0%     10.000 ± 0%   +25.00% (p=0.000 n=10)
Emit/writer/40_attrs-16   26.00 ± 0%      28.00 ± 0%    +7.69% (p=0.000 n=10)
Slog/no_attrs-16          0.000 ± 0%      0.000 ± 0%         ~ (p=1.000 n=10) ¹
Slog/3_attrs-16           0.000 ± 0%      6.000 ± 0%         ? (p=0.000 n=10)
Slog/5_attrs-16            0.00 ± 0%      10.00 ± 0%         ? (p=0.000 n=10)
Slog/10_attrs-16          5.000 ± 0%     21.000 ± 0%  +320.00% (p=0.000 n=10)
Slog/40_attrs-16          8.000 ± 0%     81.000 ± 0%  +912.50% (p=0.000 n=10)
Logr/no_attrs-16          0.000 ± 0%      0.000 ± 0%         ~ (p=1.000 n=10) ¹
Logr/3_attrs-16           4.000 ± 0%     10.000 ± 0%  +150.00% (p=0.000 n=10)
Logr/5_attrs-16           5.000 ± 0%     15.000 ± 0%  +200.00% (p=0.000 n=10)
Logr/10_attrs-16          13.00 ± 0%      29.00 ± 0%  +123.08% (p=0.000 n=10)
Logr/40_attrs-16          40.00 ± 0%     113.00 ± 0%  +182.50% (p=0.000 n=10)
geomean                              ²                ?                       ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Additional heap allocations from a dump creating using `go test -c -gcflags "-m=2" &> opts.txt && GODEBUG=allocfreetrace=1 ./benchmark.test -test.run=TestWriterLogger &> allocs.txt`:


opts.txt:
```
./writer_logger_test.go:30:22: ... argument escapes to heap:
./writer_logger_test.go:30:22:   flow: {heap} = &{storage for ... argument}:
./writer_logger_test.go:30:22:     from ... argument (spill) at ./writer_logger_test.go:30:22
./writer_logger_test.go:30:22:     from l.WithAttributes(... argument...) (call parameter) at ./writer_logger_test.go:30:22

./writer_logger_test.go:100:9: l escapes to heap:
./writer_logger_test.go:100:9:   flow: ~r0 = &{storage for l}:
./writer_logger_test.go:100:9:     from l (spill) at ./writer_logger_test.go:100:9
./writer_logger_test.go:100:9:     from return l (return) at ./writer_logger_test.go:100:2
./writer_logger_test.go:89:10: l escapes to heap:
./writer_logger_test.go:89:10:   flow: ~r0 = &{storage for l}:
./writer_logger_test.go:89:10:     from l (spill) at ./writer_logger_test.go:89:10
./writer_logger_test.go:89:10:     from return l (return) at ./writer_logger_test.go:89:3
```

allocs.txt:
```
tracealloc(0xc00017a000, 0x100, [4]attribute.KeyValue)
goroutine 4 [running]:
runtime.mallocgc(0x100, 0x5c4040, 0x1)
	/usr/local/go/src/runtime/malloc.go:1225 +0x707 fp=0xc000079b78 sp=0xc000079b10 pc=0x40eb27
runtime.newobject(0x5db1a0?)
	/usr/local/go/src/runtime/malloc.go:1324 +0x25 fp=0xc000079ba0 sp=0xc000079b78 pc=0x40eda5
go.opentelemetry.io/otel/log/benchmark.TestWriterLogger(0x0?)
	/home/rpajak/repos/opentelemetry-go/log/benchmark/writer_logger_test.go:30 +0x4be fp=0xc000079f70 sp=0xc000079ba0 pc=0x5a9c7e
testing.tRunner(0xc0001664e0, 0x62b328)
	/usr/local/go/src/testing/testing.go:1595 +0xff fp=0xc000079fc0 sp=0xc000079f70 pc=0x4d71bf
testing.(*T).Run.func1()
	/usr/local/go/src/testing/testing.go:1648 +0x25 fp=0xc000079fe0 sp=0xc000079fc0 pc=0x4d8145
runtime.goexit()
	/usr/local/go/src/runtime/asm_amd64.s:1650 +0x1 fp=0xc000079fe8 sp=0xc000079fe0 pc=0x46d4c1
created by testing.(*T).Run in goroutine 1
	/usr/local/go/src/testing/testing.go:1648 +0x3ad

tracealloc(0xc000004300, 0x180, benchmark.writerLogger)
goroutine 4 [running]:
runtime.mallocgc(0x180, 0x5f2420, 0x1)
	/usr/local/go/src/runtime/malloc.go:1225 +0x707 fp=0xc00013b598 sp=0xc00013b530 pc=0x40eb27
runtime.convT(0x5f2420, 0xc0000081a0?)
	/usr/local/go/src/runtime/iface.go:332 +0x2e fp=0xc00013b5d0 sp=0xc00013b598 pc=0x40c56e
go.opentelemetry.io/otel/log/benchmark.writerLogger.WithAttributes({{0x0, 0x0}, {0x661980, 0xc0000182a0}, {{{0x602257, 0x6}, {0x4, 0x0, {...}, {...}}}, ...}, ...}, ...)
	/home/rpajak/repos/opentelemetry-go/log/benchmark/writer_logger_test.go:89 +0x378 fp=0xc00013b878 sp=0xc00013b5d0 pc=0x5aa2b8
go.opentelemetry.io/otel/log/benchmark.(*writerLogger).WithAttributes(0x5db1a0?, {0xc00017a000?, 0x43a89e?, 0x45a851?})
	<autogenerated>:1 +0x87 fp=0xc00013bba0 sp=0xc00013b878 pc=0x5ab3e7
go.opentelemetry.io/otel/log/benchmark.TestWriterLogger(0x0?)
	/home/rpajak/repos/opentelemetry-go/log/benchmark/writer_logger_test.go:30 +0x665 fp=0xc00013bf70 sp=0xc00013bba0 pc=0x5a9e25
testing.tRunner(0xc0001664e0, 0x62b328)
	/usr/local/go/src/testing/testing.go:1595 +0xff fp=0xc00013bfc0 sp=0xc00013bf70 pc=0x4d71bf
testing.(*T).Run.func1()
	/usr/local/go/src/testing/testing.go:1648 +0x25 fp=0xc00013bfe0 sp=0xc00013bfc0 pc=0x4d8145
runtime.goexit()
	/usr/local/go/src/runtime/asm_amd64.s:1650 +0x1 fp=0xc00013bfe8 sp=0xc00013bfe0 pc=0x46d4c1
created by testing.(*T).Run in goroutine 1
	/usr/local/go/src/testing/testing.go:1648 +0x3ad
```

As far I understand this caused by the fact that `WithAttribute` is called on an interface and thus its reference arguments (slice) are being escaped. 

From https://landontclipp.github.io/blog/2023/07/15/analyzing-go-heap-escapes/#use-of-interfaces:

>  It turns out that the Go compiler is incapable of knowing at compile-time whether the underlying type in an interface could cause the reference to escape the stack. From the perspective of the function taking an interface as an argument, this knowledge is difficult to know at compile time.

If we make sure that the method is not called on an interface then the slice would not escape

```go
	l = l.WithAttributes(
		attribute.String("string", testString),
		attribute.Float64("float", testFloat),
		attribute.Int("int", testInt),
		attribute.Bool("bool", testBool),
	).(writerLogger)
```

```
./writer_logger_test.go:30:22: ... argument does not escape
```

The same is described in:  https://youtu.be/tC4Jt3i62ns?si=TaJVJsNh-pH37FTI&t=961

**Still it is worth to note that, the Bridge API user should not assume that any particular implementation is being used. It would be against the whole idea of having API separated from the SDK.**

Moreover, the returned logger in `WithAttribute` escapes, I think it is because the value is returned as an interface which is a reference type:

```
./writer_logger_test.go:100:9: l escapes to heap:
./writer_logger_test.go:100:9:   flow: ~r0 = &{storage for l}:
./writer_logger_test.go:100:9:     from l (spill) at ./writer_logger_test.go:100:9
./writer_logger_test.go:100:9:     from return l (return) at ./writer_logger_test.go:100:2
./writer_logger_test.go:89:10: l escapes to heap:
./writer_logger_test.go:89:10:   flow: ~r0 = &{storage for l}:
./writer_logger_test.go:89:10:     from l (spill) at ./writer_logger_test.go:89:10
./writer_logger_test.go:89:10:     from return l (return) at ./writer_logger_test.go:89:3
```