@pellared 2023-12-11
The version from this PR has worse performance.

The problem is that we need to "deep clone" the slices produced by each `Logger.WithAttributes` call as we need to make sure that the user does not append the same slice as the logger can be used in different places. This is done to not have problems like this one: https://blog.allegro.tech/2017/07/golang-slices-gotcha.html.

In the original design this is solved by having a separate `Clone` method on `Record`. Should we then also create `Logger.Clone`? I think such design would be very awkward and would violate the Principle of Least Astonishment. Any other ideas?