@pellared 2023-12-18
Proposal rejected.